Improved exception handling
The exception from response method behavior changed to match the fields
that are returned from the API. This actually means changing oslo files
which is not a recommended thing to do, but I see no other way to inject a
custom error parsing to their code. Monkey-patching is worse option I
think.

Change-Id: I33b926b9541c3e16b256d7e7383fa65a89bb5642