remove or adopt incubated oslo code
Switch to oslo.i18n fully.

Move the old incubated apiclient package to storyboardclient._apiclient
to make it clear it is no longer maintained by the Oslo team.

Remove the configuration file for syncing from the Oslo incubator.

Update the flake8 and coverage settings to ignore the apiclient in its
new home until it is cleaned up and passes completely.

Story: 2000776
Change-Id: I017e965353a20e0af151f0db9dc0ea8da9ff4b2f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>