remove sphinx-isms from README.rst and clean up docs
The :doc: role is only known to Sphinx, so when we publish the
README.rst to PyPI and include that role it causes the content to not
render as proper HTML.

This patch removes some of the content from the readme and moves it to
the main index page for the documentation. It then removes the
readme.rst file from the doc build because we don't need to include
the same content multiple times.

Change-Id: Ic46c7a9d8e91f01b5e6cef5d377a56856b586c70
Signed-off-by: Doug Hellmann <doug@doughellmann.com>